Implement --max switch

+++ services/analyse_commit.rb
@@ -11,13 +11,18 @@ class AnalyseCommit
   end
 
   def call
+    return false unless needs_update?
+
     LOG.info "Analysing commit #{commit.commit_hash} (#{percent_done})..."
 
     switched = false
+    analysed = false
 
-    3.times do    # Rather than defining tool dependencies, just run all tools many times
+    # Rather than defining tool dependencies, just run all tools many times
+    # (e.g. CountTodos depends on CountFiles to create commit.files)
+    3.times do
       COMMIT_ANALYSERS.each do |tool|
-        instance = tool.new(commit: commit)
+        instance = tool.new(commit: commit, options: options)
         if instance.needs_update?
           unless switched
             execute_command "cd #{root_path} && git reset --hard && git checkout #{commit.commit_hash} && git reset --hard #{commit.commit_hash}"
@@ -25,9 +30,18 @@ class AnalyseCommit
           end
 
           LOG.info ">> #{tool}"
-          instance.call
+          result = instance.call
+          analysed ||= !!result
         end
       end
+    end
+
+    return analysed
+  end
+
+  def needs_update?
+    return COMMIT_ANALYSERS.any? do |tool|
+      tool.new(commit: commit, options: options).needs_update?
     end
   end
